split crosslink/geo
