+++ src/code/auxiliar.ts
@@ -14,16 +14,7 @@ import { faPalette } from "@fortawesome/free-solid-svg-icons/faPalette";
 import { faAsterisk } from "@fortawesome/free-solid-svg-icons/faAsterisk";
 import { faDesktop } from "@fortawesome/free-solid-svg-icons/faDesktop";
 
-//** This function generates a unique ID for an object */
-export function generateId(len = 40) {
-  const arr = new Uint8Array(len / 2);
-  window.crypto.getRandomValues(arr);
-  return Array.from(arr, (dec) => {
-    return ("0" + dec.toString(16)).substr(-2);
-  }).join("");
-}
-
-export function newColors(incoming: string) {
+function newColors(incoming: string) {
   switch (incoming) {
     case "groupa":
       return "orange";
