+++ src/code/model/task.ts
@@ -1,4 +1,4 @@
-import { generateId } from "../auxiliar";
+import { generateId } from "./utils";
 
 export type TaskState = "pending" | "assigned" | "acknowledged" | "completed";
