+++ src/code/model/utils.ts
@@ -1,0 +1,8 @@
+//** This function generates a unique ID for an object */
+export function generateId(len = 40) {
+  const arr = new Uint8Array(len / 2);
+  window.crypto.getRandomValues(arr);
+  return Array.from(arr, (dec) => {
+    return ("0" + dec.toString(16)).substr(-2);
+  }).join("");
+}
